% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summerize_DEGs.R
\name{BinarizeCSIMatrix}
\alias{BinarizeCSIMatrix}
\title{Binarizing the connection specificity index (CSI) matrix.}
\usage{
BinarizeCSIMatrix(csiMat, cutoff = 0.8)
}
\arguments{
\item{csiMat}{a CSI matrix}

\item{cutoff}{threshold for binarization. Default: 0.8}
}
\value{
a binarized CSI matrix
}
\description{
Binarizing the connection specificity index (CSI) matrix.
}
